Isolated my test case in folder
Due to the changes in 1.13 in golang, flags are failing in tests because they aren't initialized before the test begins and parse isn't happening when it should. I will look into this further but i suspect i can just add a parse flag in the test before the tests begin but seems like a work around.